packages/client_search/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "wealth_management_portal_client_search" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/packages/client_search/html'
 --cov-report xml:'../../coverage/packages/client_search/coverage.xml'
 --html='../../reports/packages/client_search/unittests/html/index.html'
 --junitxml='../../reports/packages/client_search/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "wealth_management_portal_client_search" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "wealth_management_portal.client_search"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "wealth_management_portal.common_auth",
  "aws-opentelemetry-distro==0.15.0",
  "bedrock-agentcore==0.1.7",
  "fastapi==0.128.5",
  "boto3==1.42.44",
  "strands-agents==1.25.0",
  "uvicorn==0.40.0",
  "pydantic>=2.12.5"
]

[tool.uv.sources]
"wealth_management_portal.common_auth" = { workspace = true }

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
 "fastapi[standard]==0.128.5",
 "pytest>=9.0.2",
 "pytest-cov>=7.0.0",
 "pytest-html>=4.2.0",
 "ruff>=0.15.5",
]
